feat(rewards): add LoyaltyService with atomic point awarding and points_transactions ledger
- Add Firebase.PointsTransactions enum to FirebaseKeys
- Add AppLog.rewards logger category
- Create LoyaltyService.shared singleton (no @MainActor, safe for Firestore transactions)
- Award points atomically: card.points increment + ledger write in one runTransaction
- Detect milestone crossings post-transaction; grant $20 wallet credit per milestone

+++ CoffeeCraft/CoffeeCraft/Repository/Product/FirestoreProductRepository.swift
@@ -82,7 +82,7 @@ struct FirestoreProductRepository: ProductRepositoryProtocol {
             category: data[Firebase.Products.category] as? String ?? "Others",
             available: data[Firebase.Products.available] as? Bool ?? true,
             customizations: data[Firebase.Products.customizations] as? [String: [String: Double]],
-            pointsValue: data[Firebase.Products.pointsValue] as? Int ?? 0,
+            pointsValue: data[Firebase.Products.pointsValue] as? Double ?? 0,
             avgRating: data[Firebase.Products.avgRating] as? Double,
             ratingCount: data[Firebase.Products.ratingCount] as? Int,
             ratingDistribution: data[Firebase.Products.ratingDistribution] as? [String: Int]
